requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
cssselect==1.2.0
dnspython==2.7.0
idna==3.10
jmespath==1.0.1
lxml==5.3.0
packaging==24.2
parsel==1.9.1
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
ObjectId